Update addRun( ) method
The addRun() method works as expected when called, until the runNotes String is reached. The line was reading as an empty line and skipping over allowing user input. I added an extra input line to ensure that the newLine() is recognized. This may lead to problems when trying to store these notes in the runNotes variable. I will find out soon enough