Update...everything
I dug myself into a hole trying to add an array and ended up changing almost everything. It seems to work now, and I have a working while loop which is pretty cool. If I can get this Array list to work, the rest of the app should be tracking to run smoothly. Yes, that was intentional